golub {RankProd}	R Documentation

sub set of the Gene expression dataset from Golub et al. (1999)

Description

Gene expression data (500 genes and 38 tumor mRNA samples) from the leukemia microarray study of Golub et al. (1999). Original data set contain 3051 genes

Usage


data(golub)

Value

golub 	 matrix of gene expression levels for the 38 tumor mRNA samples, rows correspond to genes (500 genes) and columns to mRNA samples.

golub.cl 	 numeric vector indicating the tumor class, 27 acute lymphoblastic leukemia (ALL) cases (code 0) and 11 acute myeloid leukemia (AML) cases (code 1).

golub.gnames 	 a matrix containing the names of the 500 genes for the expression matrix golub. The three columns correspond to the gene index, ID, and Name, respectively.

Source

Golub et al. (1999). Molecular classification of cancer: class discovery and class prediction by gene expression monitoring, Science, Vol. 286:531-537.
http://www-genome.wi.mit.edu/MPR/ .

References

S. Dudoit, J. Fridlyand, and T. P. Speed (2002). Comparison of discrimination methods for the classification of tumors using gene expression data. Journal of the American Statistical Association, Vol. 97, No. 457, p. 77–87.

[Package RankProd version 2.12.0 Index]