plotRP {RankProd}	R Documentation

Graphical Display of the Rank Product/Sum analysis

Description

Plot a graphical display of the estimated pfp vs number of identified genes

Usage


     plotRP(x, cutoff=NULL)

Arguments

x 	 the value returned by function RP, RPadvance or RSadvance

cutoff 	 threshold in pfp used to select genes

Value

A graphical display of the estimated pfp vs number of identified genes, which is also the gene rank of its original rank product/sum across all comparison. If cutoff is sepcified, a horizontal line will be plotted on the graphic to indicate the positon of the cutoff point, and all genes identified will be marked red.
Two plots will be displayed, one for the identification of up-regulated genes in class 2, one for the identification of down-regulated genes in class 2

Author(s)

Fangxin Hong fhong@salk.edu

See Also

topGene RP RPadvance RSadvance

Examples


     
      # Load the data of Golub et al. (1999). data(golub) 
      #contains a 3051x38 gene expression
      # matrix called golub, a vector of length called golub.cl 
      #that consists of the 38 class labels,
      # and a matrix called golub.gnames whose third column contains the gene names.
      data(golub)
 
      #use a subset of data as example, apply the rank product method
      subset <- c(1:4,28:30)
      #Setting rand=123, to make the results reproducible,

      #identify genes that are up-regulated in class 2 
      #(class label =1)
      RP.out <- RP(golub[,subset],golub.cl[subset], rand=123)
      
      #plot the results
      plotRP(RP.out,cutoff=0.05)
      

[Package RankProd version 2.12.0 Index]